\name{airline}
\alias{airline}
\docType{data}
\title{Classical Airline Passenger Data}
\description{Monthly international airline passengers (in 1000s) from January 1949-December 1960. Series G in Box, Jenkings, and Reinsel text}
\usage{data("airline")}
\format{
  The format is:
 num [1:144] 112 118 132 129 121 135 148 148 136 119 ...
}
\source{"Time Series Analysis: Forecasting and Control" by Box, Jenkins, and Reinsel}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(airline)}
\keyword{datasets}
