\name{airlog}
\alias{airlog}
\docType{data}
\title{Natural log of airline data}
\description{Natural log of monthly international airline passengers (in 1000s) from January 1949-December 1960.   Series G in Box, Jenkings, and Reinsel text}
\usage{data("airlog")}
\format{
  The format is:
 num [1:144] 4.72 4.77 4.88 4.86 4.8 ...
}
\source{"Time Series Analysis: Forecasting and Control" by Box, Jenkins, and Reinsel}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(airlog)}
\keyword{datasets}
