\name{appy}
\alias{appy}
\docType{data}
\title{Non-perforated appendicitis  data shown in Figure 10.8 (solid line) in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Annual non-perforated appendicitis rates  for years 1970-2005}
\usage{data("appy")}
\format{
  The format is:
 num [1:36] 14.8 13.7 14.3 14.2 13 ...
}
\source{Alder, et al. (2010)Archives of Surgery 145, 63-71}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(appy)}
\keyword{datasets}
