\name{bat}
\alias{bat}
\docType{data}
\title{Bat echolocation signal shown in Figure 13.11a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Bat echolocation signal of a big brown bat}
\usage{data("bat")}
\format{
  The format is:
 num [1:381] -0.0049 -0.0083 0.0127 0.0068 -0.0259 0.0059 0.0386 -0.0405 -0.0269 0.0474 ...
}
\source{Al Feng, Beckman Center of the University of Illinois}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(bat)}
\keyword{datasets}
