\name{Bsales}
\alias{Bsales}
\docType{data}
\title{Toy Data Set of Business Sales Data}
\description{100 weeks of sales data with sales, TV advertising budget, Online advertising budget and the abount of a discount if any.}
\usage{data("Bsales")}
\references{The Time Series Toolkit}
\examples{data(Bsales)}
\keyword{datasets}
