\name{bumps16}
\alias{bumps16}
\docType{data}
\title{16 point bumps signal}
\description{Bumps signal from Donoho and Johnstone(1994) Biometrika 81,425-455}
\usage{data("bumps16")}
\format{
  The format is:
 num [1:16] 0.1 0.4 5.5 0.2 1.4 0.5 0.3 0.7 0.1 2.5 ...
}
\source{Donoho and Johnstone(1994) Biometrika 81,425-455}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(bumps16)}
\keyword{datasets}
