\name{bumps256}
\alias{bumps256}
\docType{data}
\title{256 point bumps signal}
\description{Bumps signal from Donoho and Johnstone(1994) Biometrika 81,425-455}
\usage{data("bumps256")}
\format{
  The format is:
 num [1:256] 0.00016 0.00017 0.000182 0.000195 0.000211 ...
}
\source{Donoho and Johnstone(1994) Biometrika 81,425-455}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(bumps256)}
\keyword{datasets}
