\name{cardiac}
\alias{cardiac}
\docType{data}
\title{classic Cardiac Mortality Data from 1970 - 1979 in LA}
\description{10 years of weekly cardiac mortality (cmort) data as well as the average weeky temperature (tempr) and the pollution in the air (parts).}
\usage{data("cardiac")}
\references{"The Time Series Toolkit""}
\examples{data(cardiac)}
\keyword{datasets}
