\name{cement}
\alias{cement}
\docType{data}
\title{Cement data shown in Figure 3.30a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Quarterly usage of metric tons (in thousands) of Portland cement used  from the first quarter of 1973 through the fourth quarter of 1993 in Australia} 
\usage{data("cement")}
\format{
  The format is:
 num [1:84] 1148 1305 1342 1452 1184 ...
}
\source{Australian Bureau of Statistics}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(cement)}
\keyword{datasets}
