\name{chirp}
\alias{chirp}
\docType{data}
\title{Chirp data shown in Figure 12.2a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{256 point linear chirp data}
\usage{data("chirp")}
\format{
  The format is:
List of 2
 $ x   : num [1:256] 1 1 0.98 0.95 0.91 0.86 0.8 0.72 0.63 0.53 ...
 $ spec: num [1:256] 0.511 0.568 0.733 0.991 1.32 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(chirp)}
\keyword{datasets}
