\name{dallastemp}
\alias{dallastemp}
\docType{data}
\title{Average Montly Temperature in Dallas from 1898 - 2020}
\description{Average Montly Temperature in Dallas from 1898 - 2020}
\usage{data("dallastemp")}
\format{
  The format is:
 num [1:1400] 77.2, 64.8 ...
}
\source{https://www.weather.gov/fwd/dmotemp}
\references{The Joy of Forecasting}
\examples{data(dallastemp)}
\keyword{datasets}
