\name{dfw.2011}
\alias{dfw.20110}
\docType{data}
\title{Dallas Ft. Worth monthly temperatures for January 2011 through December 2020}
\description{Dallas Ft. Worth monthly temperatures for January 2011 through December 2020 as a numeric vector}
\usage{data("dfw.2011")}
\format{
  The format is:
 num [1:120] 42.8 49.5 61.3 70.8 ... 
}
\source{weather.gov}
\examples{
data(dfw.2011)}
\keyword{datasets}
