\name{dfw.mon}
\alias{dfw.mon}
\docType{data}
\title{DFW Monthly Temperatures}
\description{Monthly average temperatures at Dallas Ft. Worth (in Fahrenheit) from January 1900 through December 2020}
\usage{data("dfw.mon")}
\format{ ts object consisting of monthly data from January 1900 through December 2020}
\source{https://www.weather.gov/fwd/dmotemp}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(dfw.mon)}
\keyword{datasets}