\name{dfw.yr}
\alias{dfw.yr}
\docType{data}
\title{DFW Annual Temperatures}
\description{Annual average temperatures at Dallas Ft. Worth (in Fahrenheit) from January 1900 through December 2020}
\usage{data("dfw.yr")}
\format{ ts object consisting of annual data from 1900 through 2020}
\source{https://www.weather.gov/fwd/dmotemp}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(dfw.yr)}
\keyword{datasets}