\name{doppler}
\alias{doppler}
\docType{data}
\title{Doppler Data}
\description{Generated Doppler data}
\usage{data("doppler")}
\format{
  The format is:
 num [1:2000] -0.00644 -0.01739 -0.02961 -0.04091 -0.04952 ...
}
\source{Simulated}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(doppler)}
\keyword{datasets}
