\name{doppler2}
\alias{doppler2}
\docType{data}
\title{Doppler signal in Figure 13.10 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Doppler signal with two time-varying frequencies}
\usage{data("doppler2")}
\format{
  The format is:
 num [1:200] -0.372 1.246 -1.163 0.261 -0.698 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(doppler2)}
\keyword{datasets}
