\name{dow.annual}
\alias{dow.annual}
\docType{data}
\title{DOW Annual Closing Averages}
\description{DOW Annual closing averages from 1915 through 2020}
\usage{data("dow.annual")}
\format{ ts object consisting of DOW Annual closing averages from 19155 through 2020}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(dow.annual)}
\keyword{datasets}