\name{dow.rate}
\alias{dow.rate}
\docType{data}
\title{DOW Daily Rate of Return Data}
\description{DOW daily rate of return data from October 1, 1928 to December 31, 2010}
\usage{data("dow.rate")}
\format{
  The format is:
 num [1:20656] 240 238 238 240 240 ...
}
\source{Public access}
\references{"Applied Statistics and Data Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(dow.rate)}
\keyword{datasets}
