\name{dow1000}
\alias{dow1000}
\docType{data}
\title{Dow Jones daily rate of return data  for 1000 days}
\description{Dow Jones daily rate of return for the 1000 trading days before December 31, 2010.}
\usage{data("dow1000")}
\format{
  The format is:
 num [1:1001] 240 238 238 240 240 ...
}
\source{Internet and shown in Figure 4.9, "Applied Time Series Analysis with R, 2nd edition", by Woodward, Gray and Elliott}
\examples{
data(dow1000)}
\keyword{datasets}
