\name{dow1921.2019}
\alias{dow1921.2019}
\docType{data}
\title{Annual DOW data 1921-2019}
\description{Annual DOW for the years 1921-2019}
\usage{data("dow1921.2019")}
\format{
  The format is:
 num [1:144] 4.72 4.77 4.88 4.86 4.8 ...
}
\source{Internet}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(dow1921.2019)}
\keyword{datasets}
