\name{dow1985}
\alias{dow1985}
\docType{data}
\title{Daily DOW Closing Prices 1985 through 2020}
\description{Daily DOW Closing Prices 1985 through 2020}
\usage{data("dow1985")}
\format{ ts object consisting of daily dow closing prices from 1985 through 2020}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(dow1985)}
\keyword{datasets}