\name{dow1995}
\alias{dow1995}
\docType{data}
\title{Monthly DOW 1995-2019}
\description{Monthly DOW for the years 195-2019}
\usage{data("dow1995")}
\format{
  The format is:
 num [1:144] 4.72 4.77 4.88 4.86 4.8 ...
}
\source{Internet}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(dow1995)}
\keyword{datasets}
