\name{dowjones2014}
\alias{dowjones2014}
\docType{data}
\title{Dow Jones daily averages for 2014}
\description{Daily Dow Jones averages for 2014}
\usage{data("dowjones2014")}
\format{
  The format is:
 num [1:252] 16441 16470 16425 16531 16463 ...
}

\source{Economic Data: Federal Reserve Bank of St. Louis. Website: https://research.stlouisfed.org/fred2/series/DJIA/downloaddata}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(dowjones2014)}
\keyword{datasets}
