\name{eco.cd6}
\alias{eco.cd6}
\docType{data}
\title{6-month rates}
\description{6-month rates 1/1/1991 through 4/1/2010}
\usage{data("eco.cd6")}
\format{
  The format is:
 num [1:469] 7.25 7.53 7.64 7.64 7.59 7.44 7.39 7.26 7.25 7.19 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(eco.cd6)}
\keyword{datasets}
