\name{eco.corp.bond}
\alias{eco.corp.bond}
\docType{data}
\title{Corporate bond rates}
\description{Corporate bond rates 1/1/1991 through 4/1/2010}
\usage{data("eco.corp.bond")}
\format{
  The format is:
 num [1:469] 4.61 5.22 5.69 6.04 6.06 5.91 5.43 5.04 4.89 4.26 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(eco.corp.bond)}
\keyword{datasets}
