\name{eco.mort30}
\alias{eco.mort30}
\docType{data}
\title{30 year mortgage rates}
\description{30-year mortgage rates 1/1/1991 through 4/1/2010}
\usage{data("eco.mort30")}
\format{
  The format is:
 num [1:469] 7.31 7.43 7.53 7.6 7.7 7.69 7.63 7.55 7.48 7.44 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(eco.mort30)}
\keyword{datasets}
