\name{fig1.10a}
\alias{fig1.10a}
\docType{data}
\title{Simulated data shown in Figure 1.10a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{This is the sum of the three signals in fig1.10b, fig1.10c, and fig1.10d}
\usage{data("fig1.10a")}
\format{
  The format is:
 num [1:1000] 0.0217 -0.1528 -0.3141 -0.4613 -0.5934 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig1.10a)}
\keyword{datasets}
