\name{fig1.10b}
\alias{fig1.10b}
\docType{data}
\title{Simulated data shown in Figure 1.10b in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Low frequency component of Figure 1.10a}
\usage{data("fig1.10b")}
\format{
  The format is:
 num [1:1000] 1 1 0.999 0.998 0.997 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{ data(fig1.10b)}
\keyword{datasets}
