\name{fig1.10c}
\alias{fig1.10c}
\docType{data}
\title{Simulated data in Figure 1.10c in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Middle frequencies component in Figure 1.10a}
\usage{data("fig1.10c")}
\format{
  The format is:
 num [1:1000] 0.73 0.646 0.56 0.471 0.381 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig1.10c)}
\keyword{datasets}
