\name{fig1.10d}
\alias{fig1.10d}
\docType{data}
\title{Simulated data in Figure 1.10d in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{High frequency component of Figure 1.10a}
\usage{data("fig1.10d")}
\format{
  The format is:
 num [1:1000] -1.71 -1.8 -1.87 -1.93 -1.97 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig1.10d)}
\keyword{datasets}
