\name{fig1.16a}
\alias{fig1.16a}
\docType{data}
\title{Simulated data for Figure 1.16a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{ Data containing two dominant frequencies}
\usage{data("fig1.16a")}
\format{
  The format is:
 num [1:250] -0.89 -3.209 0.929 -0.763 -1.972 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig1.16a)}
\keyword{datasets}
