\name{fig1.21a}
\alias{fig1.21a}
\docType{data}
\title{Simulated shown in Figure 1.21a of Woodward, Gray, and Elliott text}
\description{Simulated shown in Figure 1.21a of Woodward, Gray, and Elliott text. It illustrates the fact that frequency information is displayed better in the spectrum than the autocorrelations.}
\usage{data("fig1.21a")}
\format{
  The format is:
 num [1:250] -0.89 -3.209 0.929 -0.763 -1.972 ...
}
\source{Simulated by the authors of the Woodward, Gray, and Elliott text}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig1.21a)}
\keyword{datasets}
