\name{fig1.22a}
\alias{fig1.22a}
\docType{data}
\title{White noise data}
\description{Realization of length n=250 of white noise data, Figure 1.22a in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig1.22a")}
\format{
  The format is:
 num [1:250] 0.302 -0.691 -0.477 0.814 -0.267 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig1.22a)}
\keyword{datasets}
