\name{fig1.5}
\alias{fig1.5}
\docType{data}
\title{Simulated data shown in Figure 1.5 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated data from an ergodic AR(1) process}
\usage{data("fig1.5")}
\format{
  The format is:
 num [1:100] 0.739 -0.39 0.15 -0.627 0.262 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig1.5)}
\keyword{datasets}
