\name{fig10.11x}
\alias{fig10.11x}
\docType{data}
\title{Simulated data shown in Figure 10.11 (solid line) in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated unobservable AR(1) data in Example 10.11}
\usage{data("fig10.11x")}
\format{
  The format is:
 num [1:75] -0.2497 -0.0812 -0.6463 -1.7653 -2.719 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig10.11x)}
\keyword{datasets}
