\name{fig10.11y}
\alias{fig10.11y}
\docType{data}
\title{Simulated data shown in Figure 10.11 (dashed line) in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated observed AR(1) plus noise data in Example 10.11}
\usage{data("fig10.11y")}
\format{
  The format is:
 num [1:75] -0.74 0.045 -0.775 -2.944 -2.278 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig10.11y)}
\keyword{datasets}
