\name{fig10.1bond}
\alias{fig10.1bond}
\docType{data}
\title{Data for Figure 10.1b in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Moody's seasoned Aaa corporate bond rate, January 1, 1991-April1, 2010}
\usage{data("fig10.1bond")}
\format{
  The format is:
 num [1:232] 7.17 6.51 6.5 6.16 6.03 6.26 6.25 5.79 5.6 5.32 ...
}
\source{Internet}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig10.1bond)}
\keyword{datasets}
