\name{fig10.1cd}
\alias{fig10.1cd}
\docType{data}
\title{Data shown in Figure 10.1a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{6 month CD rate for January 1, 1991 - April 1, 2010}
\usage{data("fig10.1cd")}
\format{
  The format is:
 num [1:232] 9.04 8.83 8.93 8.86 8.86 9.01 9 8.75 8.61 8.55 ...
}
\source{Internet}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig10.1cd)}
\keyword{datasets}
