\name{fig10.1mort}
\alias{fig10.1mort}
\docType{data}
\title{Data shown in Figure 10.1c in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{30 year conventional mortgage rates: January 1, 1991-April1, 2010}
\usage{data("fig10.1mort")}
\format{
  The format is:
 num [1:232] 9.64 9.37 9.5 9.49 9.47 9.62 9.58 9.24 9.01 8.86 ...
}
\source{Internet}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig10.1mort)}
\keyword{datasets}
