\name{fig10.3x1}
\alias{fig10.3x1}
\docType{data}
\title{Variable X1 for the bivariate realization shown in Figure 10.3"}
\description{Variable X1 for the bivariate Var1) realization in Figure 10.3 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig10.3x1")}
\format{
  The format is:
  num [1:75] -0.0757 -0.2728 -0.8089 -2.4747 -5.9256 ...
}
\source{Simulated Var(1) data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig10.3x1)}
\keyword{datasets}
