\name{fig10.3x2}
\alias{fig10.3x2}
\docType{data}
\title{Variable X2 for the bivariate realization shown in Figure 10.3"}
\description{Variable X2 for the bivariate Var1) realization in Figure 10.3 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig10.3x2")}
\format{
  The format is:
 num [1:75] 0.646 -1.313 -0.191 -2.61 -4.925 ...
}
\source{Simulated Var(1) data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig10.3x2)}
\keyword{datasets}
