\name{fig11.12}
\alias{fig11.12}
\docType{data}
\title{Data shown in Figure 11.12a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated GATMA(1,0) data} 
\usage{data("fig11.12")}
\format{
  The format is:
 num [1:500] 2.18 -1.17 -3.13 -1.32 1.69 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig11.12)}
\keyword{datasets}
