\name{fig11.4a}
\alias{fig11.4a}
\docType{data}
\title{Data shown in Figure 11.4a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated FARMA(2,0) data}
\usage{data("fig11.4a")}
\format{
  The format is:
 num [1:100] 1.361 -0.369 0.881 2.362 0.236 ...
}
\source{simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig11.4a)}
\keyword{datasets}
