\name{fig12.1a}
\alias{fig12.1a}
\docType{data}
\title{Simulated data with two frequencies shown in Figure 12.1a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated two-frequency data in which the two frequencies are separated in time}
\usage{data("fig12.1a")}
\format{
  The format is:
 num [1:200] -1.22 -6.06 -9.66 -10.14 -8.58 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig12.1a)}
\keyword{datasets}
