\name{fig12.1b}
\alias{fig12.1b}
\docType{data}
\title{Simulated data with two frequencies shown in Figure 12.1b in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated two-frequency AR(4) data}
\usage{data("fig12.1b")}
\format{
  The format is:
 num [1:256] 10.081 10.835 0.532 -5.495 1.294 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig12.1b)}
\keyword{datasets}
