\name{fig13.18a}
\alias{fig13.18a}
\docType{data}
\title{Simulated data shown  in Figure 3.18a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated AR(4) data}
\usage{data("fig13.18a")}
\format{
  The format is:
 num [1:400] 1.251 1.0019 -0.0317 -1.0167 -1.4222 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig13.18a)}
\keyword{datasets}
