\name{fig13.2c}
\alias{fig13.2c}
\docType{data}
\title{TVF data shown in Figure 13.2c in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Realization from an Euler(2) model}
\usage{data("fig13.2c")}
\format{
  The format is:
 num [1:200] -13.14 -11.03 22.06 -8.92 -16.67 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig13.2c)}
\keyword{datasets}
