\name{fig3.10d}
\alias{fig3.10d}
\docType{data}
\title{AR(2) Realization (1-.95)^2X(t)=a(t)}
\description{AR(2) Realization (1-.95)^2X(t)=a(t) plotted in Figure 3.10d in "Applied Time series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig3.10d")}
\format{
  The format is:
 num [1:100] 15.3 16.3 18.6 21.2 22.8 ...
}
\details{This realization is also used in Chapter 7 of text above for testing estimation techniques}
\source{Simulated realization}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig3.10d)}
\keyword{datasets}
