\name{fig3.16a}
\alias{fig3.16a}
\docType{data}
\title{Figure 3.16a in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{Realization from the AR(3) model in Figure 3.16a}
\usage{data("fig3.16a")}
\format{
  The format is:
 num [1:200] -0.0686 0.4304 0.4786 0.9899 3.4047 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig3.16a)}
\keyword{datasets}
