\name{fig3.18a}
\alias{fig3.18a}
\docType{data}
\title{Figure 3.18a in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{Realization from the AR(3) model in Figure 3.18a}
\usage{data("fig3.18a")}
\format{
  The format is:
 num [1:200] -0.573 -0.837 -1.16 1.078 -0.561 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig3.18a)}
\keyword{datasets}
