\name{fig3.24a}
\alias{fig3.24a}
\docType{data}
\title{ARMA(2,1) realization}
\description{ARMA(2,1) realization of length n=200 phi(1)=1.6,phi(2)=-.9,theta(1)=.8 (using Box-Jenkins-Reinsel notation)}
\usage{data("fig3.24a")}
\format{
  The format is:
 num [1:200] 0.685 -1.234 -0.714 0.796 -0.96 ...
}
\source{Simulated data}
\references{Fig3.24a in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig3.24a)}
\keyword{datasets}
