\name{fig3.29a}
\alias{fig3.29a}
\docType{data}
\title{Simulated data shown in Figure 3.29a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated data from stationary seasonal model}
\usage{data("fig3.29a")}
\format{
  The format is:
 num [1:20] -7.23 -6.99 -6.9 -6.26 -3.79 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig3.29a)}
\keyword{datasets}
