\name{fig4.8a}
\alias{fig4.8a}
\docType{data}
\title{Gaussian White Noise}
\description{Gaussian White Noise, n=1000 shown in Figure 4.8a in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig4.8a")}
\format{
  The format is:
 num [1:1000] -0.585 0.177 0.284 -0.271 0.126 ...
}
\source{Simulated data}
\references{Plotted in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig4.8a)}
\keyword{datasets}
