\name{fig5.3c}
\alias{fig5.3c}
\docType{data}
\title{Data from Figure 5.3c in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{Realization of length 200 from the AR(3) model (1-.995B)(1-1.2B+.8B^2)X(t)=a(t)}
\usage{data("fig5.3c")}
\format{
  The format is:
 num [1:200] -0.503 -0.811 -0.188 1.34 2.982 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{
data(fig5.3c)}
\keyword{datasets}
