\name{fig6.11a}
\alias{fig6.11a}
\docType{data}
\title{Cyclical Data}
\description{First 50 points of data in Figure 6.11a, Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\usage{data("fig6.11a")}
\format{
  The format is:
 num [1:50] -0.682 0.15 2.262 3.079 4.122 ...
}
\source{Simulated}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig6.11a)}
\keyword{datasets}
