\name{fig6.1nf}
\alias{fig6.1nf}
\docType{data}
\title{Data in Figure 6.1 without the forecasts}
\description{Realization from the AR(1) model (1-.8B)(X(t)-25)=a(t) in Figure 6.2 and also shown in Table 6.1 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig6.1nf")}
\format{
  The format is:
 num [1:80] 25.1 27.1 27.3 25.7 23.9 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.1nf)}
\keyword{datasets}
