\name{fig6.2nf}
\alias{fig6.2nf}
\docType{data}
\title{Data in Figure 6.2 without the forecasts}
\description{Realization from the ARMA(2,1) model (1-1.2B+.6B^2)(X(t)-50)=(1-.5B)a(t) in Figure 6.2 and also shown in Table 6.1 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig6.2nf")}
\format{
  The format is:
 num [1:25] 49.5 51.1 50 49.7 50.4 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.2nf)}
\keyword{datasets}
