\name{fig6.5nf}
\alias{fig6.5nf}
\docType{data}
\title{Data in Figure 6.5 without the forecasts}
\description{Realization from the ARIMA(0,1,0) model for realization in Figure 6.5 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig6.5nf")}
\format{
  The format is:
 num [1:50] 105 104 103 102 102 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.5nf)}
\keyword{datasets}
