\name{fig6.6nf}
\alias{fig6.6nf}
\docType{data}
\title{Data in Figure 6.6 without the forecasts}
\description{Realization from the ARIMA(1,1,0) model (1-.8B)(1-B)X(t)=a(t) for realization in Figure 6.6 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig6.6nf")}
\format{
  The format is:
 num [1:50] 139 138 138 140 141 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.6nf)}
\keyword{datasets}
