\name{fig6.7nf}
\alias{fig6.7nf}
\docType{data}
\title{Data in Figure 6.2 without the forecasts}
\description{Realization from the ARIMA(0,2,0) model for realization in Figure 6.7 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("fig6.7nf")}
\format{
  The format is:
 num [1:50] -582 -579 -578 -578 -579 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.7nf)}
\keyword{datasets}
