\name{fig6.8nf}
\alias{fig6.8nf}
\docType{data}
\title{Simulated seasonal data with s=12}
\description{Simulated seasonal data designed for showing seasonal forecasts}
\usage{data("fig6.8nf")}
\format{
  The format is:
 num [1:48] 5.8 13.66 9.83 7.33 6.96 ...
}
\source{Simulated Data}
\references{"Applied Time series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(fig6.8nf)}
\keyword{datasets}
