\name{fig8.11a}
\alias{fig8.11a}
\docType{data}
\title{Data for Figure 8.11a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Realization of length n=200 from the model (1-B)(1-1.79B+1.75B^2-1.61B^3+.765B^4)X(t)=a(t)}
\usage{data("fig8.11a")}
\format{
  The format is:
 num [1:200] 83.2 80.9 78.9 80.4 85.4 ...
}
\source{Simulated data}
\references{Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig8.11a)}
\keyword{datasets}
