\name{fig8.4a}
\alias{fig8.4a}
\docType{data}
\title{Data for Figure 8.4a in Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Realization of length n=200 from the model (1-.8B)(1-1.6B+.995B^2)X(t)=a(t)}
\usage{data("fig8.4a")}
\format{
  The format is:
 num [1:200] 13.45 -5.52 -19 -21.26 -13.63 ...
}
\source{simulated data}
\references{Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig8.4a)}
\keyword{datasets}
