\name{fig8.6a}
\alias{fig8.6a}
\docType{data}
\title{Data for Figure 8.6a in Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{The realization of length n=200 is from the model (1-B)^2(1-1.2B+.6B^2)X(t)=a(t)}
\usage{data("fig8.6a")}
\format{
  The format is:
 num [1:200] 354 368 383 399 417 ...
}
\source{Simulated data}
\references{Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig8.6a)}
\keyword{datasets}
