\name{fig8.8a}
\alias{fig8.8a}
\docType{data}
\title{Data for Figure 8.8a in Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Realization of length n=200 from the model (1-B^12)(1-1.25B+.9B^2)(X(t)-50)=a(t)}
\usage{data("fig8.8a")}
\format{
  The format is:
 num [1:200] 48.9 42.9 49.3 57.3 55.5 ...
}
\source{Simulated data}
\references{Applied time series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(fig8.8a)}
\keyword{datasets}
