\name{flu}
\alias{flu}
\docType{data}
\title{Influenza data shown in Figure 10.8 (dotted line)}
\description{Annual influenza rate for years 1970-2005}
\usage{data("flu")}
\format{
  The format is:
 num [1:36] 9.75 5.82 10.99 10.41 8.42 ...
}
\source{Alder, et al. (2010)Archives of Surgery 145, 63-71}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(flu)}
\keyword{datasets}
