\name{freeze}
\alias{freeze}
\docType{data}
\title{Minimum temperature data}
\description{Each data value represents the minimum temperature over 10-day period at a location in South America}
\usage{data("freeze")}
\format{
  The format is:
 num [1:500] 8.2 12.3 9.2 8.4 10 8.8 6.8 4.8 5.2 1.7 ...
}
\source{Unknown}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(freeze)}
\keyword{datasets}
