\name{freight}
\alias{freight}
\docType{data}
\title{Freight data}
\description{9 years of monthly freight shipment data}
\usage{data("freight")}
\format{
  The format is:
 num [1:120] 1299 1148 1345 1363 1374 ...
}
\source{Unknown}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(freight)}
\keyword{datasets}
