\name{global.temp}
\alias{global.temp}
\docType{data}
\title{Global Temperature Data: 1850-2009}
\description{Annual temperature anomalies from the average for the years 1850-2009}
\usage{data("global.temp")}
\format{
  The format is:
List of 2
 $ year  : num [1:160] 1850 1851 1852 1853 1854 ...
 $ annual: num [1:160] -0.447 -0.292 -0.294 -0.337 -0.307 -0.321 -0.406 -0.503 -0.513 -0.349 ...
}
\source{Climatic Research Unit at East Anglia, England, in conjunction with the Met Office Hadley Centre}

\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(global.temp)}
\keyword{datasets}

