\name{global2020}
\alias{global2020}
\docType{data}
\title{Global Temperature Data: 1880-2009}
\description{Annual temperature anomalies from the average for the years 1850-2009}
\usage{data("global.temp")}
\format{
  The format is: ts file containing annual temperatures from 1880 through 2020
}
\source{ncdc.noaa.gov}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(global2020)}
\keyword{datasets}

