\name{hadley}
\alias{hadley}
\docType{data}
\title{Global temperature data}
\description{Global temperature data for 1850-2009. The data are temperature anomalies, i.e. departures from the average for 1850-2009}
\usage{data("hadley")}
\format{
  The format is:
 num [1:160] -0.447 -0.292 -0.294 -0.337 -0.307 -0.321 -0.406 -0.503 -0.513 -0.349 ...
}
\source{Met Office Hadley Centre}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(hadley)}
\keyword{datasets}
