\name{kingkong}
\alias{kingkong}
\docType{data}
\title{King Kong Eats Grass}
\description{Digitized record taken at 8,000 Hz of voltage readings obtained from the acoustical energy generated by Wayne Woodward speaking the words "King Kong eats grass" while a fan was blowing in the background}
\usage{data("kingkong")}
\format{
  The format is:
 num [1:15418] -0.001831 -0.000916 -0.003357 -0.002716 -0.000977 ...
}
\source{See description above}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(kingkong)}
\keyword{datasets}
