\name{lavon}
\alias{lavon}
\docType{data}
\title{Lavon lake water levels}
\description{Data given in feet above sea level. Quarterly data, 1982-2009}
\usage{data("lavon")}
\format{
  The format is:
 num [1:112] 495 492 500 491 492 ...
}
\source{http://lavon.uslakes.info/levelcal.asp}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(lavon)}
\keyword{datasets}
