\name{lavon15}
\alias{lavon15}
\docType{data}
\title{Lavon Lake Levels to September 30, 2015}
\description{Feet above sea level for Lavon Lake, quarterly data through September 2015. An extension of data lavon}
\usage{data("lavon15")}
\format{
  The format is:
 num [1:135] 495 492 500 491 492 ...
}
\source{Lake Data internet}
\examples{data(lavon15)}
\keyword{datasets}
