\name{linearchirp}
\alias{linearchirp}
\docType{data}
\title{Linear chirp data.}
\description{256 point linear chirp data,  the first 150 points of which are shown in Figure 3.16(a) Time Series Analysis for Data Science: Analysis and Forecasting by Woodward, Sadler, and Robertson}
\usage{data("linearchirp")}
\format{
  The format is:
List of 2
 $ x   : num [1:256] 1 1 0.98 0.95 0.91 0.86 0.8 0.72 0.63 0.53 ...
 $ spec: num [1:256] 0.511 0.568 0.733 0.991 1.32 ...
}
\source{Simulated data}
\references{Time Series Analysis for Data Science: Analysis and Forecasting by Woodward, Sadler, and Robertson}
\examples{data(linearchirp)}
\keyword{datasets}
