\name{llynx}
\alias{llynx}
\docType{data}
\title{Log (base 10) of lynx data}
\description{The log (base 10) of the annual number of lynx trapped in the Mackenzie River district of the North-West Canada (dataset lynx in this package)}
\usage{data("llynx")}
\format{
  The format is:
 Time-Series [1:114] from 1821 to 1934: 2.43 2.51 2.77 2.94 3.17 ...
}
\source{Tong (1977). Journal of the Royal Statistical Society A, 432-436.}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(llynx)}
\keyword{datasets}
