\name{lynx}
\alias{lynx}
\docType{data}
\title{Lynx data}
\description{The lynx data are the annual number of lynx trapped in the Mackenzie River district of Canada}
\usage{data("lynx")}
\format{
  The format is:
 Time-Series [1:114] from 1821 to 1934: 269 321 585 871 1475 ...
}
\source{Tong (1977). Journal of the Royal Statistical Society A, 432-436.}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(lynx)}
\keyword{datasets}
