\name{ma2.table7.1}
\alias{ma2.table7.1}
\docType{data}
\title{Simulated MA(2) data}
\description{This realization is used to obtain the innovations estimates shown in Table 7.1}
\usage{data("ma2.table7.1")}
\format{
  The format is:
 num [1:400] 1.299 1.831 -0.162 -0.648 1.243 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(ma2.table7.1)}
\keyword{datasets}
