\name{mass.mountain}
\alias{mass.mountain}
\docType{data}
\title{Massachusettts Mountain Earthquake Data}
\description{Lg wave from from an earthquake known as Massachusetts Mountain Earthquake(5 August 1971), which was recorded at the Mina Nevada station}
\usage{data("mass.mountain")}
\format{
  The format is:
 num [1:454] -0.03655 -0.01774 0.00218 0.01193 0.00915 ...
}
\source{Gupta, Chan, and Wagner (2005). Regional sources discrimination of small events based on the use of Lg wavetrain, Bulletin of the Seismological Society of America 95, 341-346.}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(mass.mountain)}
\keyword{datasets}
