\name{MedDays}
\alias{MedDays}
\docType{data}
\title{Median days a house stayed on the market}
\description{Median days a house stayed on the market between July 2016 and April 2020}
\usage{data("MedDays")}
\format{ ts object consisting of monthly data from July 2016 through April 2020}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(MedDays)}
\keyword{datasets}