\name{melanoma2.0}
\alias{melanoma2.0}
\docType{data}
\title{The Number of Melanomas per Year per 100,000 in Connecticut versus Sunspots Per Year Between 1939 and 1972}
\description{The Number of Melanomas per Year per 100,000 in Connecticut versus Sunspots Per Year Between 1939 and 1972.  The number of sunspots is the is from SILSO and reflects the new counting method of sunpsots that was implemented in 2015.  Note that this is thus a different data than was originally used by Houghton et al in 1978}
\usage{data("sunspot2.0")}
\format{
  The format is:
ts object: start 1700, end = 2020, 8.3  18.3  26.7  38.3  60.0  96.7  48.3  ... }
\source{Source for Sunspots: WDC-SILSO, Royal Observatory of Belgium, Brussels http://www.sidc.be/silso/datafiles
Source for Melanoma: Houghton, A., E. W. Munster, and M. V. Viola. (1978). Increased Incidence of Malignant Melanoma After Peaks of Sunspot Activity. The Lancet, 8, 759–760.}
\references{Time Series for Data Science}
\examples{data(melanoma2.0)}
\keyword{datasets}
