\name{mm.eq}
\alias{mm.eq}
\docType{data}
\title{Massachusetts Mountain Earthquake data shown in Figure 13.13a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Lg wave for Massachusetts Mountain Earthquake}
\usage{data("mm.eq")}
\format{
  The format is:
 num [1:454] -0.03655 -0.01774 0.00218 0.01193 0.00915 ...
}
\source{Gupta, et al. (2005) Bulletin of the Seismological Society of America 95, 341-346.}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(mm.eq)}
\keyword{datasets}
