\name{NAICS}
\alias{NAICS}
\docType{data}
\title{Monthly Retail Sales Data}
\description{Monthly sales for the North American Industry Classification System (NAICS) code 44X72: Retail Trade and Food Services: 1992-2019}
\usage{data("NAICS")}
\format{ ts object consisting of monthly data from January 1992- December 2019}
\source{https://www.weather.gov/fwd/dmotemp}
\references{"Kaggle" and "US Census Bureau" websites}
\examples{data(NAICS)}
\keyword{datasets}