\name{nbumps256}
\alias{nbumps256}
\docType{data}
\title{256 noisy bumps signal}
\description{Noisy bumps signal shown in Figure 12.11(a) in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\usage{data("nbumps256")}
\format{
  The format is:
 num [1:256] -0.234 0.123 0.303 0.134 -0.513 ...
}
\source{Donoho and Johnstone(1994) Biometrika 81,425-455}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(nbumps256)}
\keyword{datasets}
