\name{nile.min}
\alias{nile.min}
\docType{data}
\title{Annual minimal water levels of Nile river}
\description{Water levels for 622 through 1284 measured at Roda gauge near Cairo (Tousson, 1925)}
\usage{data("nile.min")}
\format{
  The format is:
 Time-Series [1:663] from 622 to 1284: 1157 1088 1169 1169 984 ...
}
\source{Tousson, O. (1925) M\'emoire sur l'Histoire du Nil, Volume 18 in M\'emoires a l'Institut d'Egypte, pp. 366-404.}
\references{Beran, J. (1994) Statistics for Long-Memory Processes, Chapman Hall: Englewood, NJ.}
\examples{data(nile.min)}
\keyword{datasets}
