\name{noctula}
\alias{noctula}
\docType{data}
\title{Nyctalus noctula echolocation data}
\description{Echolocation signal for the Nyctalus noctula hunting bat}
\usage{data("noctula")}
\format{
  The format is:
 num [1:96] -18 16 -5 -17 21 -6 -17 20 -6 -16 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\keyword{datasets}
