\name{NSA}
\alias{NSA}
\docType{data}
\title{Monthly Total Vehicle Sales}
\description{Monthly Total Vehicle Sales (TOTALNSA)
in the United States from January 1976 - December 2019}
\usage{data("NSA")}
\format{ ts object consisting of monthly data from January 1976- December 2019}
\source{https://www.weather.gov/fwd/dmotemp}
\references{"Kaggle" and "US Census Bureau" websites}
\examples{data(NSA)}
\keyword{datasets}