\name{ozona}
\alias{ozona}
\docType{data}
\title{Number of Chicken Fried Steaks Sold at Ozona in Dallas in June and July 2019}
\description{Number of Chicken Fried Steaks Sold at Ozona in Dallas in June and July 2019}
\usage{data("ozona")}
\format{
  The format is:
 num [1:381] -0.0049 -0.0083 0.0127 0.0068 -0.0259 0.0059 0.0386 -0.0405 -0.0269 0.0474 ...
}
\source{Cory Wauson, GM of Ozona}
\references{The Joy of Forcasting}
\examples{data(ozona)}
\keyword{datasets}
