\name{patemp}
\alias{patemp}
\docType{data}
\title{Pennsylvania average monthly temperatures}
\description{Pennsylvania average monthly temperatures}
\usage{data("patemp")}
\format{
  The format is:
 num [1:180] 38.1 38.3 44.5 52.3 59.2 70.6 73.9 71.3 63.9 57.3 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and elliott}
\examples{data(patemp)}
\keyword{datasets}
