\name{prob10.4}
\alias{prob10.4}
\docType{data}
\title{Data matrix for Problem 10.4 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{Matrix containing a bivariate VAR data set}
\usage{data("prob10.4")}
\format{
  The format is:
 num [1:100, 1:2] 0 0.7184 -0.3448 -2.1638 -0.0342 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2] "X1" "X2"
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob10.4)}
\keyword{datasets}
