\name{prob10.7x}
\alias{prob10.7x}
\docType{data}
\title{Data for Problem 10.7 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{This realization is the same unobservable data as in prob10.6x}
\usage{data("prob10.7x")}
\format{
  The format is:
 num [1:9] 2.61 0.69 0.64 0.37 -0.79 -1.63 -1.14 -1.2 -3.13
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob10.7x)}
\keyword{datasets}
