\name{prob10.7y}
\alias{prob10.7y}
\docType{data}
\title{Simulated observed data for Problem 10.6 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Kalman filter example data}
\usage{data("prob10.7y")}
\format{
  The format is:
 num [1:9] 3.28 -0.05 0.64 0.31 -0.9 -2.4 -1.83 -1.93 -3.52
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob10.7y)}
\keyword{datasets}