\name{prob11.5}
\alias{prob11.5}
\docType{data}
\title{Data for Problem 11.5 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated fractional long memory data}
\usage{data("prob11.5")}
\format{
  The format is:
 num [1:10] 4.2 -2.5 8.4 14.6 7 9.6 19.8 4.8 6.5 8.3
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob11.5)}
\keyword{datasets}
