\name{prob12.1c}
\alias{prob12.1c}
\docType{data}
\title{Data for Problem 12.1c and 12.3c in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Data from a problem set in the wavelet chapter}
\usage{data("prob12.1c")}
\format{
  The format is:
 num [1:200] 9.49 8.01 3.43 -1.85 -4.99 -7.21 -5.61 -2.34 2.16 3.88 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob12.1c)}
\keyword{datasets}
