\name{prob12.3a}
\alias{prob12.3a}
\docType{data}
\title{Data for Problem 12.3a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Data from a problem set in the wavelet chapter}
\usage{data("prob12.3a")}
\format{
  The format is:
 num [1:512] -3.09 8.43 -9.74 8.44 -3.46 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob12.3a)}
\keyword{datasets}
