\name{prob12.3b}
\alias{prob12.3b}
\docType{data}
\title{Data for Problem 12.3b in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Data from a problem set in the wavelet chapter}
\usage{data("prob12.3b")}
\format{
  The format is:
 num [1:256] 1 1 1 1 1 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob12.3b)}
\keyword{datasets}
