\name{prob12.6c}
\alias{prob12.6c}
\docType{data}
\title{Data set for Problem 12.6(C) in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated TVF data set}
\usage{data("prob12.6c")}
\format{
  The format is:
 num [1:512] -0.482 -0.569 -0.656 -0.743 -0.83 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob12.6c)}
\keyword{datasets}
