\name{prob13.2}
\alias{prob13.2}
\docType{data}
\title{Data for Problem 13.2 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Simulated data from  cosine-plus-noise model}
\usage{data("prob13.2")}
\format{
  The format is:
 num [1:256] 1.524 5.886 5.939 4.319 0.573 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(prob13.2)}
\keyword{datasets}
