\name{prob8.1a}
\alias{prob8.1a}
\docType{data}
\title{Data for Problem 8.1 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{See title above}
\usage{data("prob8.1a")}
\format{
  The format is:
 num [1:200] 2.19 0.48 0.06 3.86 3.6 -3.38 6.23 1.95 1.4 -5.35 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob8.1a)}
\keyword{datasets}
