\name{prob8.1b}
\alias{prob8.1b}
\docType{data}
\title{Data for Problem 8.1 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{See title above}
\usage{data("prob8.1b")}
\format{
  The format is:
 num [1:200] 1.54 -0.13 1.93 0.29 -0.13 -0.23 1.27 1.01 -0.65 1.68 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob8.1b)}
\keyword{datasets}
