\name{prob8.1c}
\alias{prob8.1c}
\docType{data}
\title{Data for Problem 8.1 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{See title above}
\usage{data("prob8.1c")}
\format{
  The format is:
 num [1:200] 0.33 -0.53 -2.36 2.48 -0.36 -2.02 1.87 -0.73 0.41 2.41 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob8.1c)}
\keyword{datasets}
