\name{prob8.1d}
\alias{prob8.1d}
\docType{data}
\title{Data for Problem 8.1 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\description{See title above}
\usage{data("prob8.1d")}
\format{
  The format is:
 num [1:200] -0.07 -1.74 -1.37 -0.52 0.14 0.07 -1.5 1.88 -0.03 -1.81 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob8.1d)}
\keyword{datasets}
