\name{prob9.6c1}
\alias{prob9.6c1}
\docType{data}
\title{Data set 1 for Problem 6.1c}
\description{Data set 1 for Problem 6.1c in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott. It is either from line plus noise or random walk with drift.}
\usage{data("prob9.6c1")}
\format{
  The format is:
 num [1:100] -0.2924 0.0206 0.6595 0.3819 0.0269 ...
}
\source{Simulated data}

\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob9.6c1)}
\keyword{datasets}
