\name{prob9.6c2}
\alias{prob9.6c2}
\docType{data}
\title{Data set 2 for Problem 6.1c} 
\description{Data set 2 for Problem 6.1c in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott. It is either from line plus noise or random walk with drift.}
\usage{data("prob9.6c2")}
\format{
  The format is:
 num [1:100] -0.925 -2.679 -2.378 -3.03 -2.157 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott.}
\examples{data(prob9.6c2)}
\keyword{datasets}
