\name{prob9.6c3}
\alias{prob9.6c3}
\docType{data}
\title{Data set 3 for Problem 6.1c}
\description{Data set 3 for Problem 6.1c in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott. It is either from line plus noise or random walk with drift.}
\usage{data("prob9.6c3")}
\format{
  The format is:
 num [1:100] -2.79 -3.32 -3.51 -5.13 -3.51 ...
}
\source{Simulated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob9.6c3)}
\keyword{datasets}
