\name{prob9.6c4}
\alias{prob9.6c4}
\docType{data}
\title{Data set 4 for Problem 6.1c}
\description{Data set 4 for Problem 6.1c in "Applied Time Series and Data Analysis with R, 2nd edition" by Woodward, Gray, and Elliott. It is either from line plus noise or random walk with drift.}
\usage{data("prob9.6c4")}
\format{
  The format is:
 num [1:100] -0.0599 -0.0214 0.6589 -0.151 0.4043 ...
}
\source{Simulated data}
\references{"Applied Time Series and Data Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(prob9.6c4)}
\keyword{datasets}
