\name{rate}
\alias{rate}
\docType{data}
\title{Daily DOW rate of Return}
\description{Daily DOW rate of return from 1971 through 2020}
\usage{data("rate")}
\format{ ts object consisting of daily dow rate of return from 1971 through 2020}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(rate)}
\keyword{datasets}