\name{rates}
\alias{rates}
\docType{data}
\title{Daily DOW Rates of Return}
\description{Daily DOW rates of return from 1971 through 2020}
\usage{data("rates")}
\format{ ts object consisting of daily dow rates of return from 1971 through 2020}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(rates)}
\keyword{datasets}