\name{SB.Ocean.Temp.MKO}
\alias{SB.Ocean.Temp.MKO}
\docType{data}
\title{Average Daily Ocean Temperature taken from a sensor at the MKO site in Santa Barbara California from October 25, 2018 to November 5th, 2018 }
\description{Average Daily Ocean Temperature taken from a sensor at the MKO site in Santa Barbara California from October 25, 2018 to November 5th, 2018.  The data set was created by Santa Barbara Coastal LTER}
\usage{data("SB.Ocean.Temp.MKO")}
\source{Santa Barbara Coastal LTER, Gretchen E Hofmann, and Libe Washburn. 2020. SBC LTER: Ocean: Time-series: nearshore calibrated pH and temperature outside of reefs, ongoing since 2011. urn:node:LTER. https://pasta.lternet.edu/package/metadata/eml/knb-lter-sbc/6005/1.}
\references{The Joy of Forecasting}
\examples{data(SB.Ocean.Temp.MKO)}
\keyword{datasets}
