\name{ss08.1850}
\alias{ss08.1850}
\docType{data}
\title{Sunspot data from 1850 through 2008 for matching with global temperature data (hadley)}
\description{Sunspot data from 1850 through 2008 for matching with global temperature data (hadley) for purposes of testing for association in Example 10.5 of "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\usage{data("ss08.1850")}
\format{
  The format is:
 num [1:160] 66.6 64.5 54.1 39 20.6 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(ss08.1850)}
\keyword{datasets}
