\name{ss08}
\alias{ss08}
\docType{data}
\title{Sunspot Data}
\description{Annual average sunspot numbers for the years 1749-2008}
\usage{data("ss08")}
\format{
  The format is:
 num [1:260] 80.9 83.4 47.7 47.8 30.7 ...
}
\source{Internet-open source}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(ss08)}
\keyword{datasets}

