\name{starwort.ex}
\alias{starwort.ex}
\docType{data}
\title{Starwort Explosion data shown in Figure 13.13a in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{Lg wave for Starwort explosion data}
\usage{data("starwort.ex")}
\format{
  The format is:
 num [1:420] 43245 48408 47565 7372 -62277 ...
}
\source{Gupta, et al. (2005) Bulletin of the Seismological Society of America 95, 341-346.}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(starwort.ex)}
\keyword{datasets}
