\name{sunspot.classic}
\alias{sunspot.classic}
\docType{data}
\title{Classic Sunspot Data: 1749-1924}
\description{The classic 176 point sunspot data from 1749-1924 that has been widely modeled}
\usage{data("sunspot.classic")}
\format{
  The format is:
 num [1:176] 80.9 83.4 47.7 47.8 30.7 12.2 9.6 10.2 32.4 47.6 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(sunspot.classic)}
\keyword{datasets}
