\name{sunspot.new}
\alias{sunspot.new}
\docType{data}
\title{The Number of Sunspots Per Year Between 1700 and 2020}
\description{The Number of Sunspots Per Year Between 1700 and 2020}
\usage{data("sunspot.new")}
\format{
  The format is:
ts object: start 1700, end = 2020, 8.3  18.3  26.7  38.3  60.0  96.7  48.3  ... }
\source{Source: WDC-SILSO, Royal Observatory of Belgium, Brussels http://www.sidc.be/silso/datafiles}
\references{The Joy of Forecasting}
\examples{data(sunspot.new)}
\keyword{datasets}
