\name{sunspot2.0.month}
\alias{sunspot2.0.month}
\docType{data}
\title{Monthly Sunspot2.0 Numbers}
\description{Monthly sunspot2.0 numbers from January 1749 through December 2020}
\usage{data("sunspot2.0.month")}
\format{ ts object consisting of monthly data from January 1749 through December 2020}
\source{https://www.sidc.oma.be/silso}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(sunspot2.0.month)}
\keyword{datasets}