\name{table10.1.noise}
\alias{table10.1.noise}
\docType{data}
\title{Noise related to data set, the first 5 points of which are shown in Table 10.1 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{The data in Table 10.1 are of the form Y(t)=X(t)+n(t).  This data set contains the values for n(t).}
\usage{data("table10.1.noise")}
\format{
  The format is:
 num [1:75] -0.49 0.126 -0.129 -1.179 0.441 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(table10.1.noise)}
\keyword{datasets}
