\name{table10.1.signal}
\alias{table10.1.signal}
\docType{data}
\title{Underlying, unobservable signal (X(t), the first 5 points of which are shown in Table 10.1 in Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\description{The X(t) data is unobservable, and is a realization from an AR(1) model}
\usage{data("table10.1.signal")}
\format{
  The format is:
 num [1:75] -0.2497 -0.0812 -0.6463 -1.7653 -2.719 ...
}
\source{Simulated data}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(table10.1.signal)}
\keyword{datasets}
