\name{table7.1}
\alias{table7.1}
\docType{data}
\title{MA(2) data for Table 7.1}
\description{MA(2) data for Table 7.1 in "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott.  Uses function ia in package itsmr to show steps in the innovations algorithm for estimating the MA parameters and white noise variance}
\usage{data("table7.1")}
\format{
  The format is:
 num [1:400] 0.4481 0.5497 -1.6586 -3.1653 -0.0314 ...
}
\source{Generated data}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{data(table7.1)}
\keyword{datasets}
