\name{TaxiImpute}
\alias{TaxiImpute}
\docType{data}
\title{NYC Taxi Data for Imputation Example}
\description{This is the the NYC Yellow Taxi trips data (as a ts object) with March 2020 left as an NA in order to demonstrate a simple method of imputing a data value (linear interpolation)}
\usage{data("TaxiImpute")}
\source{https://www1.nyc.gov/site/tlc/about/aggregated-reports.page}
\references{"The Joy of Forecasting" by Woodward, Gray, and Elliott}
\examples{data(TaxiImpute)}
\keyword{datasets}

