\name{tesla}
\alias{tesla}
\docType{data}
\title{Tesla Stock Prices}
\description{Teslas daily stock prices from January 1, 2020 through April 30, 2021}
\usage{data("tesla")}
\format{ ts object consisting of daily adjusted close price for TSLA from January 1, 2020 through April 30, 2021}
\source{https://finance.yahoo.com}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(tesla)}
\keyword{datasets}