\name{tswge-package}
\alias{tswge-package}
\alias{tswge}
\docType{package}
\title{Time Series package for Woodward, Gray, and Elliott text}
\description{These functions and data sets accompany the book "Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\author{Wayne Woodward <waynew@smu.edu> }
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\examples{
data(wages)
plotts.wge(wages) 
}
