\name{tx.unemp.adj}
\alias{tx.unemp.adj}
\docType{data}
\title{Texas Seasonally Adjusted Unnemployment Rates}
\description{Monthly seasonally adjusted unemployment rate in Texas for the years 2000-2019}
\usage{data("tx.unemp.adj")}
\format{ ts object consisting of monthly seasonally adjusted unemployment rate from January 2000 through December 2019}
\source{https://twc.texas.gov}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(tx.unemp.adj)}
\keyword{datasets}