\name{tx.unemp.unadj}
\alias{tx.unemp.unadj}
\docType{data}
\title{Texas Unadjusted Unnemployment Rates}
\description{Monthly unemployment rate in Texas for the years 2000-2019}
\usage{data("tx.unemp.unadj")}
\format{ ts object consisting of monthly unadjusted unemployment rate from January 2000 through December 2019}
\source{https://twc.texas.gov}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(tx.unemp.unadj)}
\keyword{datasets}