\name{us.retail}
\alias{us.retail}
\docType{data}
\title{Quarterly US Retail Sales}
\description{Quarterly US retail sales (in $millions) from the fourth quarter of 1999 through the second quarter of 2021}
\usage{data("us.retail")}
\format{ ts object consisting of quarterly US retail sales (in $millions) from the fourth quarter of 1999 through the second quarter of 2021}
\source{https://www.fred.stlouis.org}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(us.retail)}
\keyword{datasets}