\name{uspop}
\alias{uspop}
\docType{data}
\title{Annual US Population}
\description{Annual US population from 1900 through 2020}
\usage{data("uspop")}
\format{ ts object consisting of annual data from 1900 through 2020}
\source{https://www.usafacts.org}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(uspop)}
\keyword{datasets}