\name{wages}
\alias{wages}
\docType{data}
\title{Daily wages in Pounds from 1260 to 1944 for England}
\description{This data set contains the average English daily wages in pounds for each year from 1260 to 1944, inclusive.}
\usage{data("wages")}
\format{
  The format is:
 num [1:735] 4.41 4.63 4.38 4.52 4.42 4.64 4.44 5.15 5.23 4.42 ...
}
\source{Data Market Time Series Data Library (citing: Makridakis, Wheelwright and Hyndman (1998))}
\examples{data(wages)}
\keyword{datasets}
