\name{whale}
\alias{whale}
\docType{data}
\title{Whale click data}
\description{256 point whale click echolocation signal}
\usage{data("whale")}
\format{
  The format is:
 num [1:286] 0.0014 -0.008 0.01126 0.00412 0.0069 ...
}
\source{Stan Kuczaj from University of Southern Mississippi}
\references{Applied Time Series Analysis with R, second edition by Woodward, Gray, and Elliott}
\examples{data(whale)}
\keyword{datasets}
