\name{wtcrude2020}
\alias{wtcrude2020}
\docType{data}
\title{Monthly WTI Crude Oil Prices}
\description{Monthly WTI crude oil prices from January 1990 through December 2020}
\usage{data("wtcrude2020")}
\format{ ts object consisting of monthly data from January 1990 through December 2020}
\source{https://fred.stlouis.org}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(wtcrude2020)}
\keyword{datasets}