\name{wtcrude2021}
\alias{wtcrude2021}
\docType{data}
\title{Average Monthly West Texas Intermediate Crude Oil Price from Jan 1990 to Jan 2021}
\description{Average Monthly West Texas Intermediate Crude Oil Price from Jan 1990 to Jan 2021}
\usage{data("wtcrude2021")}
\format{
  ts object starting at January 1990 and ending Jan 2021
}
\source{https://fred.stlouisfed.org/series/POILWTIUSDM}
\references{"The Joy of Forecasting"}
\examples{data(wtcrude)}
\keyword{datasets}
