\name{yellowcab.precleaned}
\alias{yellowcab.precleaned}
\docType{data}
\title{Yellow Cab Trips Per Month}
\description{Number of trips oper month for Yellow Cabs in NY City from January 2019 through December 2020}
\usage{data("yellowcab.precleaned")}
\format{ ts object consisting of monthly number of trips from January 2019 through December 2020}
\source{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(yellowcab.precleaned)}
\keyword{datasets}