\name{bitcoin}
\alias{bitcoin}
\docType{data}
\title{Daily Bitcoin Prices From May 1, 2020 to April 30, 2021}
\description{This dataset contains the daily price of bitcoin from May 1, 2021 to April 30, 2021. The data was gathered from Yahoo Finance on April 30, 2020 and included missing values on October 9, 12 and 13 of 2020.  Yahoo Finance has since filled in the correct values which can be compared with the imputed values described in the book.}
\usage{data("bitcoin")}
\format{
  The format is:
 num [1:461] 7200.174 6985.470 7344.884  ...
}
\source{Yahoo Finance}
\references{"Practical Time Series for Data Scientiests by Woodward, Sadler and Robertson"}
\examples{data(bitcoin)}
\keyword{datasets}
