\name{cardiac}
\alias{cardiac}
\docType{data}
\title{Weekly Cardiac Mortality Data}
\description{Weekly cardiac mortality, temperatures, and pollution measures for the years 1970-1978}
\usage{data("cardiac")}
\format{ ts object consisting of weekly data}
\source{Shumway and Stoffer, 1999)}
\references{"Time Series for Data Sience:Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(cardiac)}
\keyword{datasets}