\name{dallastemp}
\alias{dallastemp}
\docType{data}
\title{Average Montly Temperature in Dallas from 2006 - 2020}
\description{Average Montly Temperature in Dallas from 2006 - 2020}
\usage{data("dallastemp")}
\format{
  The format is:
 num [1:180] 55.1 49.4 62.2 72.0 78.0 83.7 87.7 89.8 ...
}
\source{https://www.weather.gov/fwd/dmotemp}
\references{The Joy of Forecasting}
\examples{data(dallastemp)}
\keyword{datasets}
