\name{dfw.2011}
\alias{dfw.2011}
\docType{data}
\title{DFW Monthly Temperatures from January 2011 through December 2020}
\description{Monthly average temperatures at Dallas Ft. Worth (in Fahrenheit) from January 2011 through December 2020}
\usage{data("dfw.2011")}
\format{ ts object consisting of monthly data from January 1900 trough December 2020}
\source{https://www.weather.gov/fwd/dmotemp}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(dfw.2011)}
\keyword{datasets}