\name{NY_Taxi}
\alias{NY_Taxi}
\docType{data}
\title{Number of Yellow Taxi Trips in NYC from Jan 2019 to Feb 2021}
\description{This is the number of Yellow Taxi trips in NYC from January 2019 to Februarty 2021}
\usage{data("NY_Taxi")}
\source{https://www1.nyc.gov/site/tlc/about/aggregated-reports.page}
\references{The Joy of Forecasting}
\examples{data(NY_Taxi)}
\keyword{datasets}
