\name{ozona}
\alias{ozona}
\docType{data}
\title{Daily Number of Chicken-Fried Steaks Sold}
\description{Daily number of chicken-fried steaks sold at Ozona Bar and Grill during June and July 2019}
\usage{data("ozona")}
\format{ ts object consisting of number of chicken fried steaks sold daily during June and July, 2019}
\source{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(ozona)}
\keyword{datasets}