\name{plotts.wge}
\alias{plotts.wge}
\title{Plot a time series realization}
\description{Given a realization contained in a vector, this function plots it as a time series realization}
\usage{plotts.wge(x)}
\arguments{
  \item{x}{The vector containing the time series realization to be plotted}
}
\value{Simply a plot of the realization}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\author{Wayne Woodward}
\examples{plotts.wge(rnorm(100))}

\keyword{ Plot }
\keyword{ Realization}
