\name{SB.Ocean.Temp.MKO}
\alias{SB.Ocean.Temp.MKO}
\docType{data}
\title{Average Daily Ocean Temperature taken from a sensor at the MKO site in Santa Barbara California from October 25, 2018 to November 5th, 2018 }
\description{Average Daily Ocean Temperature taken from a sensor at the MKO site in Santa Barbara California from October 25, 2018 to November 5th, 2018.  The data set was created by Santa Barbara Coastal LTER}
\usage{data("SB.Ocean.Temp.MKO")}
\source{https://search.dataone.org/view/https%3A%2F%2Fpasta.lternet.edu%2Fpackage%2Fmetadata%2Feml%2Fknb-lter-sbc%2F6005%2F1}
\references{The Joy of Forecasting}
\examples{data(SB.Ocean.Temp.MKO)}
\keyword{datasets}
