\name{sunspot2.0}
\alias{sunspot2.0}
\docType{data}
\title{Annual Sunspot2.0 Numbers}
\description{Annual sunspot2.0 numbers from 1700 through 2020}
\usage{data("sunspot2.0")}
\format{ ts object consisting of annual data from 1700 through 2020}
\source{https://www.sidc.oma.be/silso}
\references{"Time Series for Data Sience: Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\examples{data(sunspot2.0)}
\keyword{datasets}