\name{wtcrude}
\alias{wtcrude}
\docType{data}
\title{West Texas Intermediate Crude Oil Prices}
\description{Monthly West Texas intermediate crude oil prices from January 2000 through October 2009.}
\usage{data("wtcrude")}
\format{
  The format is:
 num [1:118] 27.2 29.4 29.9 25.7 28.8 ...
}
\source{Internet}
\references{"Applied Time Series Analysis with R, 2nd edition" by Woodward, Gray, and Elliott}
\keyword{datasets}
