\name{uspop}
\alias{uspop}
\docType{data}
\title{US population}
\description{US estimated annual population from 1900 through 2020.}
\usage{data("uspop")}
\format{ts object consisting of annual data from 1700 through 2020}
\source{Internet}
\references{"Time Series for Data Science:  Analysis and Forecasting" by Woodward, Sadler, and Robertson}
\keyword{datasets}
