\name{yellowcab.precleaned}
\alias{yellowcab.precleaned}
\docType{data}
\title{Precleaned Yellow Cab data}
\description{The number of Yellow Cab Trips in NYC before and during the COVID outbreak:  January 2019 through February 2021}
\usage{data("yellowcab.precleaned")}
\format{
  The format is:
 Time-Series [1:26] from 2019 to 2021: 247315 250654 252634 247742 ...
}
\source{NYC Taxi and Limousine website}
\references{Time Series for Data Science Woodward, Sadler, and Robertson}
\examples{data(yellowcab.precleaned)}
\keyword{datasets}
